Update the contract to target API structure.
The target API structure uses a numeric PIN to identify each product uniquely and lookup for them using the PIN. The PIN generator is already established and since the product data set may not be fully known at the time of generation, the contract has been updated to use the PIN as the unique ID as well. Hash of the product details set is still calculated to support future product validation, subset of product details is available in the contract. A product data update has been added to support target platform update operation.